prove: Different implementation for GetProofSubSet
GetProofSubSet now asks for the positions the caller wants proven in the
proof instead of the positions to be removed from the proof.  This
allows for a new guarantee where the returned hashes are in the same
order as the wanted positions.

The algorithm is also changed to rely on the simpler calculateHashes()
instead of the more complicated RemoveTargets().